+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 ### Fixed
 - Fix enablingSidebar including section and nested
 - Fix duplicate sidebar when creating nested
+- Fix sidebar sorting
 
 ---